% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labelled.R
\name{extract_labelled}
\alias{extract_labelled}
\alias{[.labelled}
\alias{[<-.labelled}
\title{Extract or Replace Parts of Labelled Vectors}
\usage{
\method{[}{labelled}(x, i, ...)

\method{[}{labelled}(x, i, ...) <- value
}
\arguments{
\item{x}{An atomic vector inheriting the \code{labelled} class.}

\item{i}{The elements to extract.}

\item{...}{Arguments to pass to other methods.}

\item{value}{typically a vector of similar class of length \code{i}}
}
\description{
Extraction and replacement methods for labelled vectors.
}
\examples{

x <- set_label(1:10, "Integers")
x[1:3]

x[3] <- pi
x

}
\seealso{
\code{\link{Extract}}
}
