% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_labelled.R
\name{is_labelled}
\alias{is_labelled}
\alias{is.labelled}
\title{Evaluate if a vector is labelled}
\usage{
is.labelled(x)

is_labelled(x)
}
\arguments{
\item{x}{An atomic vector}
}
\value{
Returns a \code{logical(1)}.
}
\description{
Functions to determine if a vector has a label.
}
\section{Functional Requirements}{

\enumerate{
 \item Return a logical value of length 1.
 \item Cast an error if \code{x} is not atomic.
}
}

