% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.labelled.R
\name{print.labelled}
\alias{print.labelled}
\title{Print Method for Labelled Vectors}
\usage{
\method{print}{labelled}(x, ...)
}
\arguments{
\item{x}{A vector inheriting class \code{labelled}}

\item{...}{Additional arguments to pass to other methods.}
}
\description{
Labelled vectors are printed with their label appearing above
  the content of the vector.
}
